calibrate_water_linreg

This function uses NEON validation data to apply drift corrections to measured ambient water isotope ratios. In brief, ambient water isotope ratios are calibrated by generating regressions using reference water measurements bracketing an ambient period. Three reference waters are measured once per day, with several injections per reference water. Due to memory effects, only the last three are used currently to generate calibration equations. Regressions between measured d18O and d2H values and NEON-provisioned known reference values are generated, and used to calibrate the period of ambient measurements between them if the r2 of the regression is greater than a threshold value (by default, this is 0.95). Most of this function deals with selecting the appropriate calibration data and determining calibration quality. This function also contains a wrapper for calibrate_ambient_water_linreg, which calibrates the ambient water data using the calibration parameters generated in this function. This function also copies over data in the qfqm and ucrt hdf5 data groups.

Usage

calibrate_water_linreg_bymonth(
  inname,
  outname,
  site,
  time_diff_betweeen_standards = 1800,
  filter_data = TRUE,
  force_cal_to_beginning = TRUE,
  force_cal_to_end = TRUE,
  r2_thres = 0.95
)

Arguments

inname

Name of the input file.

outname

Name of the output file.

site

Four-letter NEON code for site being processed.

time_diff_betweeen_standards

Time (in seconds) required between consecutive standard measurements.

filter_data

Apply median absolute deviation filter from Brock 86 to remove impulse spikes?

force_cal_to_beginning

Extend first calibration to the beginning of the file?

force_cal_to_end

Extend last calibration to the end of the file?

r2_thres

Minimum r2 threshold of an "acceptable" calibration. Acts to remove calibration periods where a measurement error makes relationship nonlinear. Default = 0.95

Value

nothing to the workspace, but creates a new output file of calibrated carbon isotope data.

Details

IMPORTANT NOTE Currently this function does not apply a correction for humidity dependence of Picarro isotopic measurements. This is because the data to implement these corrections is not yet publicly available. Caution is suggested when analyzing data at low humidities, below ~5000 ppm, with likely higher biases at lower humidity values.

Author

Rich Fiorella rfiorella@lanl.gov